% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findZOfData.R
\name{findzOfData}
\alias{findzOfData}
\title{Create z from vector}
\usage{
findzOfData(x)
}
\arguments{
\item{x}{quantitative vector}
}
\value{
a list containing z and x
}
\description{
Create z from vector
}
\examples{
findzOfData(1:4)
}
