% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybarplot.R
\name{mybinbarplot}
\alias{mybinbarplot}
\title{Create a barplot of random samples}
\usage{
mybinbarplot(iter, n, p)
}
\arguments{
\item{iter}{iterations through the sample}

\item{n}{sample size}

\item{p}{probability of Success}
}
\value{
barplot with table
}
\description{
Create a barplot of random samples
}
\examples{
mybinbarplot(500, 10, 1)
}
