% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{the singular dataset}
}
\value{
a confidence interval for the dataset
}
\description{
Confidence Interval
}
\examples{
myci(c(1,2,3,4,5,6,7))
}
