% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{my first function}
\usage{
myfun(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A vector of squared components
}
\description{
my first function
}
\examples{
myfun(1:10)
}
