% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Find the area of a curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the function}

\item{sigma}{the standard deviation of the function}

\item{a}{the max x-coordinate of the area}
}
\value{
area to terminal and graph of area under the curve
}
\description{
Find the area of a curve
}
\examples{
myncurve(10,5,6)
}
