% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Find optimal number of tickets to sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of available seats}

\item{gamma}{Probability the plane will truly be overbooked}

\item{p}{probability that the person will show}
}
\value{
2 graphs continuous and discrete for the optimal number of tickets sold, with a list of all the variables
}
\description{
Find optimal number of tickets to sell
}
\examples{
ntickets(200, .02, .95)
}
